Added tool tip and clean up
