+++ src/main/java/io/github/mzmine/modules/visualization/kendrickmassplot/KendrickMassPlotModule.java
@@ -19,7 +19,6 @@
 package io.github.mzmine.modules.visualization.kendrickmassplot;
 
 import java.util.Collection;
-import java.util.logging.Logger;
 import javax.annotation.Nonnull;
 import io.github.mzmine.datamodel.MZmineProject;
 import io.github.mzmine.modules.MZmineModuleCategory;
@@ -34,8 +33,6 @@ import io.github.mzmine.util.ExitCode;
  * @author Ansgar Korf (ansgar.korf@uni-muenster.de)
  */
 public class KendrickMassPlotModule implements MZmineRunnableModule {
-
-  private static final Logger logger = Logger.getLogger(KendrickMassPlotModule.class.getName());
 
   private static final String MODULE_NAME = "Kendrick mass plot";
   private static final String MODULE_DESCRIPTION = "Kendrick mass plot.";
