+++ src/main/java/io/github/mzmine/modules/visualization/kendrickmassplot/KendrickMassPlotWindowFX.fxml
@@ -4,6 +4,7 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -108,7 +109,10 @@
                            <children>
                               <Label text="Shift" GridPane.halignment="CENTER" />
                               <Label text="Charge" GridPane.columnIndex="1" GridPane.halignment="CENTER" />
-                              <Label text="Divisor" GridPane.columnIndex="2" GridPane.halignment="CENTER" />
+                              <Label text="Divisor" GridPane.columnIndex="2" GridPane.halignment="CENTER">
+                                 <tooltip>
+                                    <Tooltip fx:id="tooltipYAxisLabel" />
+                                 </tooltip></Label>
                               <Button fx:id="shiftUpYAxis" mnemonicParsing="false" onAction="#shiftUpYAxis" GridPane.halignment="CENTER" GridPane.rowIndex="1">
                                  <graphic>
                                     <ImageView pickOnBounds="true">
@@ -201,7 +205,11 @@
                            <children>
                               <Label text="Shift" GridPane.halignment="CENTER" />
                               <Label text="Charge" GridPane.columnIndex="1" GridPane.halignment="CENTER" />
-                              <Label text="Divisor" GridPane.columnIndex="2" GridPane.halignment="CENTER" />
+                              <Label text="Divisor" GridPane.columnIndex="2" GridPane.halignment="CENTER">
+                                 <tooltip>
+                                    <Tooltip fx:id="tooltipXAxisLabel" />
+                                 </tooltip>
+                              </Label>
                               <Button fx:id="shiftUpXAxis" mnemonicParsing="false" onAction="#shiftUpXAxis" GridPane.halignment="CENTER" GridPane.rowIndex="1">
                                  <graphic>
                                     <ImageView pickOnBounds="true">
@@ -297,7 +305,11 @@
                            <children>
                               <Label text="Shift" GridPane.halignment="CENTER" />
                               <Label text="Charge" GridPane.columnIndex="1" GridPane.halignment="CENTER" />
-                              <Label text="Divisor" GridPane.columnIndex="2" GridPane.halignment="CENTER" />
+                              <Label text="Divisor" GridPane.columnIndex="2" GridPane.halignment="CENTER">
+                                 <tooltip>
+                                    <Tooltip fx:id="tooltipZAxisLabel" text="Empty Tooltip" />
+                                 </tooltip>
+                              </Label>
                               <Button fx:id="shiftUpZAxis" mnemonicParsing="false" onAction="#shiftUpZAxis" GridPane.halignment="CENTER" GridPane.rowIndex="1">
                                  <graphic>
                                     <ImageView pickOnBounds="true">
